kopano-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kopano"
version = "0.1.0"
description = "Multi-AI Collaboration Platform CLI (MCP)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Kholofelo Robyn Rababalela" }]

dependencies = [
    "typer[all]>=0.12",
    "rich>=13",
    "pydantic-settings>=2",
    "python-dotenv>=1",
    "azure-monitor-opentelemetry>=1.8",
    "litellm>=1.40",
    "sqlalchemy>=2.0",
    "google-generativeai>=0.8.0",
    "openai>=1.55.0",
    "httpx>=0.27.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "websockets>=12.0",
]

[project.scripts]
kopano = "kopano.cli:app"

[tool.setuptools.packages.find]
include = ["kopano"]
exclude = ["CLI", "public", "Skimatics", "certificates"]
